requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-heroku==0.3.1
djangorestframework==3.11.0
idna==2.9
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
